+++ backfill/repository/storage_repository.go
@@ -8,18 +8,18 @@ import (
 
 type Urn struct {
 	Ilk   string
-	IlkID int `db:"ilk_id"`
+	IlkID int64 `db:"ilk_id"`
 	Urn   string
-	UrnID int `db:"urn_id"`
+	UrnID int64 `db:"urn_id"`
 }
 
 type StorageRepository interface {
 	GetOrCreateUrn(urn, ilk string) (int64, error)
-	GetUrnByID(id int) (Urn, error)
+	GetUrnByID(id int64) (Urn, error)
 	InsertDiff(diff types.RawDiff) error
-	VatIlkArtExists(ilkID, headerID int) (bool, error)
-	VatUrnArtExists(urnID, headerID int) (bool, error)
-	VatUrnInkExists(urnID, headerID int) (bool, error)
+	VatIlkArtExists(ilkID, headerID int64) (bool, error)
+	VatUrnArtExists(urnID, headerID int64) (bool, error)
+	VatUrnInkExists(urnID, headerID int64) (bool, error)
 }
 
 type storageRepository struct {
@@ -34,7 +34,7 @@ func NewStorageRepository(db *postgres.DB) StorageRepository {
 	return storageRepository{db: db}
 }
 
-func (repo storageRepository) GetUrnByID(id int) (Urn, error) {
+func (repo storageRepository) GetUrnByID(id int64) (Urn, error) {
 	var urn Urn
 	err := repo.db.Get(&urn, `
 		SELECT DISTINCT urns.id AS urn_id, ilks.ilk, ilks.id AS ilk_id, urns.identifier AS urn
@@ -52,19 +52,19 @@ func (repo storageRepository) InsertDiff(diff types.RawDiff) error {
 	return err
 }
 
-func (repo storageRepository) VatIlkArtExists(ilkID, headerID int) (bool, error) {
+func (repo storageRepository) VatIlkArtExists(ilkID, headerID int64) (bool, error) {
 	var exists bool
 	err := repo.db.Get(&exists, `SELECT EXISTS (SELECT 1 FROM maker.vat_ilk_art WHERE ilk_id = $1 and header_id = $2)`, ilkID, headerID)
 	return exists, err
 }
 
-func (repo storageRepository) VatUrnArtExists(urnID, headerID int) (bool, error) {
+func (repo storageRepository) VatUrnArtExists(urnID, headerID int64) (bool, error) {
 	var exists bool
 	err := repo.db.Get(&exists, `SELECT EXISTS (SELECT 1 FROM maker.vat_urn_art WHERE urn_id = $1 and header_id = $2)`, urnID, headerID)
 	return exists, err
 }
 
-func (repo storageRepository) VatUrnInkExists(urnID, headerID int) (bool, error) {
+func (repo storageRepository) VatUrnInkExists(urnID, headerID int64) (bool, error) {
 	var exists bool
 	err := repo.db.Get(&exists, `SELECT EXISTS (SELECT 1 FROM maker.vat_urn_ink WHERE urn_id = $1 and header_id = $2)`, urnID, headerID)
 	return exists, err
